04-airflow-pyspark-iceberg/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-apache-spark==5.2.1
pyspark==3.5.1
boto3==1.42.62
requests==2.32.5